added authorization for todos: new, index, create and destroy

+++ app/controllers/todos_controller.rb
@@ -2,14 +2,17 @@ class TodosController < ApplicationController
 
   def new
     @todo = Todo.new
+    authorize @todo
   end
 
   def index
     @todos = current_user.todos
+    authorize @todos
   end
 
   def create
     @todo = current_user.todos.create(todo_params)
+    authorize @todo
     if @todo.persisted?
       redirect_to todos_path, notice: "Your new TODO was saved"
     else
@@ -20,6 +23,7 @@ class TodosController < ApplicationController
 
   def destroy
     @todo = Todo.find params[:id]
+    authorize @todo
     if @todo.destroy
       redirect_to todos_path, notice: "Todo deleted successfully"
     else
